Delete misago.posts app
