--- misago/posts/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class MisagoPostsConfig(AppConfig):
-    name = "misago.posts"
-    label = "misago_posts"
-    verbose_name = "Misago Posts"
