--- misago/posts/hooks/__init__.py
@@ -1,3 +1,0 @@
-from .redirect_to_post import redirect_to_post_hook
-
-__all__ = ["redirect_to_post_hook"]
